pyproject.toml
[project]
name = "vibe-code-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "groq>=0.37.1",
    "langchain>=1.2.17",
    "langchain-core>=1.3.2",
    "langchain-groq>=1.1.2",
    "langgraph>=1.1.10",
    "pip>=26.1.1",
    "pydantic>=2.13.3",
    "python-dotenv>=1.2.2",
]
